test: shard tui gateway s2 clusters
