+++ tests/test_tui_gateway_server.py
@@ -3255,173 +3255,6 @@ def test_config_sync_adopts_baseline_when_agent_already_on_target(monkeypatch):
     assert session["config_model_seen"] == ("old/model", "")
 
 
-def test_config_sync_switches_when_only_provider_differs(monkeypatch):
-    _patch_config_model(monkeypatch, "old/model", provider="nous")
-    session = _sync_test_session(config_model_seen=("old/model", ""))
-    calls = []
-    monkeypatch.setattr(
-        server,
-        "_apply_model_switch",
-        lambda sid, sess, raw, **kw: calls.append(raw),
-    )
-
-    server._sync_agent_model_with_config("sid", session)
-
-    assert calls == ["old/model --provider nous"]
-
-
-def test_config_sync_failure_emits_error_once_per_edit(monkeypatch):
-    _patch_config_model(monkeypatch, "broken/model")
-    session = _sync_test_session(config_model_seen=("old/model", ""))
-
-    def boom(*a, **k):
-        raise ValueError("no such model")
-
-    monkeypatch.setattr(server, "_apply_model_switch", boom)
-    emits = []
-    monkeypatch.setattr(
-        server, "_emit", lambda ev, sid, payload: emits.append((ev, payload))
-    )
-
-    server._sync_agent_model_with_config("sid", session)
-    server._sync_agent_model_with_config("sid", session)
-
-    assert len(emits) == 1
-    assert emits[0][0] == "error"
-    assert "broken/model" in emits[0][1]["message"]
-
-
-def test_config_sync_config_wins_over_env_seed(monkeypatch):
-    # Hosted instances set HERMES_INFERENCE_MODEL as a provision-time seed;
-    # the per-turn sync must follow config.yaml edits, not stay pinned to it.
-    monkeypatch.setenv("HERMES_INFERENCE_MODEL", "seed/model")
-    monkeypatch.delenv("HERMES_MODEL", raising=False)
-    monkeypatch.setattr(server, "_load_cfg", lambda: {"model": {"default": "new/model"}})
-    session = _sync_test_session(config_model_seen=("seed/model", ""))
-    calls = []
-    monkeypatch.setattr(
-        server,
-        "_apply_model_switch",
-        lambda sid, sess, raw, **kw: calls.append(raw),
-    )
-
-    server._sync_agent_model_with_config("sid", session)
-
-    assert calls == ["new/model"]
-    assert session["config_model_seen"] == ("new/model", "")
-
-
-def test_config_sync_ignores_env_seed_without_config_model(monkeypatch):
-    # `hermes --tui -m <model>` sets HERMES_MODEL/HERMES_INFERENCE_MODEL as a
-    # launch-scoped seed. When config.yaml has NO model.default (typical
-    # custom-provider-only setup), the sync must NOT adopt the env seed as a
-    # config target — doing so replayed the -m flag as a /model switch and
-    # (with persist_switch_by_default=True) wrote it into config.yaml
-    # permanently.
-    monkeypatch.setenv("HERMES_MODEL", "one-shot/model")
-    monkeypatch.setenv("HERMES_INFERENCE_MODEL", "one-shot/model")
-    monkeypatch.setattr(
-        server, "_load_cfg", lambda: {"model": {"provider": "custom:mylocal"}}
-    )
-    session = _sync_test_session()
-    monkeypatch.setattr(
-        server,
-        "_apply_model_switch",
-        lambda *a, **k: pytest.fail("env seed must not trigger a config sync switch"),
-    )
-
-    server._sync_agent_model_with_config("sid", session)
-
-
-def test_config_model_target_never_reads_env(monkeypatch):
-    monkeypatch.setenv("HERMES_MODEL", "seed/model")
-    monkeypatch.setenv("HERMES_INFERENCE_MODEL", "seed/model")
-    monkeypatch.setattr(server, "_load_cfg", lambda: {"model": {"provider": "nous"}})
-
-    assert server._config_model_target() == ("", "nous")
-
-
-def test_apply_model_switch_persist_override_false_never_persists(monkeypatch):
-    # Internal callers (config sync, /moa one-shot + restore) pass
-    # persist_override=False; even with persist_switch_by_default=True the
-    # switch must not write config.yaml.
-    import types as _types
-
-    result = _types.SimpleNamespace(
-        success=True,
-        new_model="new/model",
-        target_provider="nous",
-        base_url="",
-        api_key="key",
-        api_mode="chat_completions",
-        warning_message="",
-        model_info=None,
-        error_message="",
-    )
-    monkeypatch.setattr(
-        "hermes_cli.model_switch.switch_model", lambda **kw: result
-    )
-    monkeypatch.setattr(
-        "hermes_cli.model_switch.resolve_persist_behavior",
-        lambda *a: pytest.fail("persist_override must bypass resolve_persist_behavior"),
-    )
-    monkeypatch.setattr(
-        server, "_persist_model_switch",
-        lambda _r: pytest.fail("persist_override=False must not persist"),
-    )
-    monkeypatch.setattr(
-        "hermes_cli.model_cost_guard.expensive_model_warning",
-        lambda *a, **k: None,
-    )
-    session = {"agent": None}
-
-    out = server._apply_model_switch(
-        "sid", session, "new/model --provider nous", persist_override=False
-    )
-
-    assert out["value"] == "new/model"
-    assert session["model_override"]["model"] == "new/model"
-
-
-def test_startup_runtime_uses_tui_provider_env(monkeypatch):
-    monkeypatch.setenv("HERMES_MODEL", "nous/hermes-test")
-    monkeypatch.setenv("HERMES_TUI_PROVIDER", "nous")
-    monkeypatch.delenv("HERMES_INFERENCE_PROVIDER", raising=False)
-
-    assert server._resolve_startup_runtime() == ("nous/hermes-test", "nous")
-
-
-def test_startup_runtime_does_not_treat_inference_provider_as_explicit(monkeypatch):
-    monkeypatch.setenv("HERMES_MODEL", "nous/hermes-test")
-    monkeypatch.delenv("HERMES_TUI_PROVIDER", raising=False)
-    monkeypatch.setenv("HERMES_INFERENCE_PROVIDER", "nous")
-    monkeypatch.setattr(
-        "hermes_cli.models.detect_static_provider_for_model",
-        lambda model, provider: None,
-    )
-
-    assert server._resolve_startup_runtime() == ("nous/hermes-test", None)
-
-
-def test_startup_runtime_detects_provider_for_model_env(monkeypatch):
-    monkeypatch.setenv("HERMES_MODEL", "sonnet")
-    monkeypatch.delenv("HERMES_TUI_PROVIDER", raising=False)
-    monkeypatch.delenv("HERMES_INFERENCE_PROVIDER", raising=False)
-    monkeypatch.setattr(server, "_load_cfg", lambda: {"model": {"provider": "auto"}})
-
-    def fake_detect(model, current_provider):
-        assert model == "sonnet"
-        assert current_provider == "auto"
-        return "anthropic", "anthropic/claude-sonnet-4.6"
-
-    monkeypatch.setattr(
-        "hermes_cli.models.detect_static_provider_for_model", fake_detect
-    )
-
-    assert server._resolve_startup_runtime() == (
-        "anthropic/claude-sonnet-4.6",
-        "anthropic",
-    )
 
 
 def test_load_fallback_model_merges_chain_providers_first(monkeypatch):
@@ -3534,41 +3367,6 @@ def test_background_agent_kwargs_preserves_empty_fallback_chain(monkeypatch):
 
     assert kwargs["fallback_model"] == []
 
-
-def test_startup_runtime_resolves_short_alias_without_network(monkeypatch):
-    monkeypatch.setenv("HERMES_MODEL", "sonnet")
-    monkeypatch.delenv("HERMES_TUI_PROVIDER", raising=False)
-    monkeypatch.delenv("HERMES_INFERENCE_PROVIDER", raising=False)
-    monkeypatch.setattr(server, "_load_cfg", lambda: {"model": {"provider": "auto"}})
-    monkeypatch.setattr(
-        "hermes_cli.models.fetch_openrouter_models",
-        lambda *_args, **_kwargs: (_ for _ in ()).throw(
-            AssertionError("network lookup should not run")
-        ),
-    )
-
-    model, provider = server._resolve_startup_runtime()
-
-    assert provider == "anthropic"
-    assert model.startswith("claude-sonnet")
-
-
-def test_startup_runtime_does_not_call_network_detector(monkeypatch):
-    monkeypatch.setenv("HERMES_MODEL", "sonnet")
-    monkeypatch.delenv("HERMES_TUI_PROVIDER", raising=False)
-    monkeypatch.delenv("HERMES_INFERENCE_PROVIDER", raising=False)
-    monkeypatch.setattr(server, "_load_cfg", lambda: {"model": {"provider": "auto"}})
-    monkeypatch.setattr(
-        "hermes_cli.models.detect_provider_for_model",
-        lambda *_args, **_kwargs: (_ for _ in ()).throw(
-            AssertionError("network detector called")
-        ),
-    )
-
-    model, provider = server._resolve_startup_runtime()
-
-    assert model
-    assert provider in {None, "anthropic"}
 
 
 def _session(agent=None, **extra):
